+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.4.1] - Unreleased
 
+### Changes
+- Split utils up and put them in an `utils` package
+
 
 ## [1.4.0] - 2018.05.18